refactor: move Point class into main directory
