+++ jarvis/src/main/java/ru/itis/ads/jarvis/JarvisAlgorithm.java
@@ -1,7 +1,5 @@
 package ru.itis.ads.jarvis;
 
-import ru.itis.ads.jarvis.point.Point;
-
 import java.util.ArrayList;
 import java.util.List;
