+++ jarvis/src/main/java/ru/itis/ads/jarvis/Point.java
@@ -1,4 +1,4 @@
-package ru.itis.ads.jarvis.point;
+package ru.itis.ads.jarvis;
 
 import java.util.Objects;
